move api-based embeddings/reranking calls to api server out of model server, added/modified unit tests
